@@ -17,9 +17,9 @@
   },
   "dependencies": {
     "autoprefixer": "8.2.0",
-    "babel-core": "6.26.0",
-    "babel-loader": "7.1.4",
-    "babel-preset-env": "1.6.1",
+    "babel-core": "7.0.0-beta.3",
+    "babel-loader": "^7.1.4",
+    "babel-preset-env": "^7.0.0-beta.3",
     "browser-sync": "2.23.6",
     "cf-atomic-component": "1.5.4",
     "cf-buttons": "4.6.0",
@@ -73,7 +73,7 @@
     "uglifyjs-webpack-plugin": "1.2.4",
     "ustream-embedapi": "1.0.0",
     "vinyl-named": "1.1.0",
-    "webpack": "4.5.0",
+    "webpack": "4.6.0",
     "webpack-stream": "4.0.3",
     "xdr": "0.5.3"
   },
@@ -82,7 +82,7 @@
     "babel-jest": "22.4.3",
     "chai": "4.1.2",
     "chai-as-promised": "7.1.1",
-    "cucumber": "4.0.0",
+    "cucumber": "4.2.1",
     "fancy-log": "1.3.2",
     "jasmine-reporters": "2.3.0",
     "jasmine-spec-reporter": "4.2.1",
@@ -93,11 +93,11 @@
     "merge-stream": "1.0.1",
     "minimist": "1.2.0",
     "mkdirp": "0.5.1",
-    "protractor": "5.3.0",
+    "protractor": "5.3.1",
     "protractor-accessibility-plugin": "github:cfpb/protractor-accessibility-plugin#on-page-load",
-    "protractor-cucumber-framework": "4.2.0",
+    "protractor-cucumber-framework": "5.0.0",
     "protractor-retry": "1.1.9",
-    "selenium-webdriver": "3.6.0",
+    "selenium-webdriver": "4.0.0-alpha.1",
     "snyk": "1.70.2",
     "wcag": "0.3.0"
   },
